Create Bookingform.cs
This codes displays how to filter through various items and selections at once in a C# DataGridView. It allow displays the selected user's details from the Search form as well as displaying any items the user decides to add to their basket in a list and updating the cost accordingly.